Push absolute path logic into driver layer
https://github.com/wix/Detox/pull/1919#issuecomment-593073179
